+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/test/Providers/ModelProviders/TestData/ModelCustomizationTests/CanChangeEnumNullableTypeModelProperty/MockInputModel.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections.Generic;
+using SampleTypeSpec;
+
+namespace Sample.Models;
+
+public partial class MockInputModel
+{
+    public Foo Prop1 { get; };
+}